requirements.txt
pytest==6.0.1
allure-pytest==2.8.18
allure-python-commons==2.8.18
requests==2.24.0